Move geo world to own file
