+++ Makefile
@@ -1,13 +1,13 @@
 CFLAGS = -g -Wshadow -std=c++11 -pedantic -Wall -Wextra -Wcast-align -Wcast-qual -Wctor-dtor-privacy -Winit-self -Wlogical-op -Wmissing-declarations -Wmissing-include-dirs -Wnoexcept  -Woverloaded-virtual -Wredundant-decls -Wsign-conversion -Wsign-promo -Wstrict-null-sentinel -Wstrict-overflow=5 -Wswitch-default -Wundef -Wno-unused -Wold-style-cast -Wformat-nonliteral -Wpointer-arith -Winline -Wwrite-strings -Wno-unused-parameter -Wfloat-equal -fsanitize=undefined -fsanitize=unreachable -fsanitize=leak -fsanitize=enum -fsanitize=builtin -fsanitize=shift-exponent -fsanitize=pointer-compare -fsanitize=pointer-subtract -fsanitize=address -Werror
 
-OBJECTS = obj/main.o obj/Construction.o obj/drawers.o obj/Geoapp.o obj/geo.o obj/geoImpl.o obj/Input.o obj/inputSFML.o obj/keyStates.o
+OBJECTS = obj/main.o obj/Construction.o obj/drawers.o obj/Geoapp.o obj/geo.o obj/geoImpl.o obj/geoWorld.o obj/Input.o obj/inputSFML.o obj/keyStates.o
 
 bin/app: $(OBJECTS)
 	g++ $(OBJECTS) -o bin/app -lsfml-window -lsfml-system -lsfml-graphics $(CFLAGS)
 
 obj/main.o: main.cpp source/Geoapp.h source/geo.h source/inputSFML.h \
  source/Input.h source/Construction.h source/drawers.h source/uiOptions.h \
- source/conditionTracker.h | obj
+ source/conditionTracker.h source/geoWorld.h | obj
 	g++ main.cpp -o obj/main.o -c $(CFLAGS)
 
 obj/Construction.o: source/Construction.cpp source/Construction.h \
@@ -19,7 +19,8 @@ obj/drawers.o: source/drawers.cpp source/drawers.h source/geo.h | obj
 
 obj/Geoapp.o: source/Geoapp.cpp source/Geoapp.h source/geo.h \
  source/inputSFML.h source/Input.h source/Construction.h source/drawers.h \
- source/uiOptions.h source/conditionTracker.h source/keyStates.h | obj
+ source/uiOptions.h source/conditionTracker.h source/geoWorld.h \
+ source/keyStates.h | obj
 	g++ source/Geoapp.cpp -o obj/Geoapp.o -c $(CFLAGS)
 
 obj/geo.o: source/geo.cpp source/geo.h | obj
@@ -27,6 +28,11 @@ obj/geo.o: source/geo.cpp source/geo.h | obj
 
 obj/geoImpl.o: source/geoImpl.cpp source/geoImpl.h source/geo.h | obj
 	g++ source/geoImpl.cpp -o obj/geoImpl.o -c $(CFLAGS)
+
+obj/geoWorld.o: source/geoWorld.cpp source/geoWorld.h source/drawers.h \
+ source/geo.h source/uiOptions.h source/conditionTracker.h \
+ source/Construction.h | obj
+	g++ source/geoWorld.cpp -o obj/geoWorld.o -c $(CFLAGS)
 
 obj/Input.o: source/Input.cpp source/Input.h | obj
 	g++ source/Input.cpp -o obj/Input.o -c $(CFLAGS)
@@ -36,7 +42,8 @@ obj/inputSFML.o: source/inputSFML.cpp source/inputSFML.h source/Input.h | obj
 
 obj/keyStates.o: source/keyStates.cpp source/keyStates.h source/Geoapp.h \
  source/geo.h source/inputSFML.h source/Input.h source/Construction.h \
- source/drawers.h source/uiOptions.h source/conditionTracker.h | obj
+ source/drawers.h source/uiOptions.h source/conditionTracker.h \
+ source/geoWorld.h | obj
 	g++ source/keyStates.cpp -o obj/keyStates.o -c $(CFLAGS)
 
 obj:
