+++ source/inputSFML.h
@@ -1,5 +1,6 @@
 /* file responsible for interpreting sfml key inputs and passing them to input class
  * */
+#pragma once
 #include "Input.h"
 #include <unordered_map>
 #include <SFML/Graphics.hpp>
