+++ source/uiOptions.h
@@ -1,6 +1,7 @@
 /* file responsible for customization of condition tracker for needs of
  * user interface
  * */
+#pragma once
 #include "geo.h"
 #include "conditionTracker.h"
 #include "Construction.h"
